move screens and components into own folders
