+++ src/components/index.js
@@ -1,0 +1,9 @@
+import { EyeIcon } from './EyeIcon';
+import { Loading } from './Loading';
+import { NavigationBar } from './NavigationBar';
+
+export {
+  EyeIcon,
+  Loading,
+  NavigationBar
+};
