+++ src/screens/index.js
@@ -1,0 +1,9 @@
+import { HomeScreen } from './HomeScreen';
+import { LogInScreen } from './LogInScreen';
+import { SignUpScreen } from './SignUpScreen';
+
+export {
+  HomeScreen,
+  LogInScreen,
+  SignUpScreen
+};
